Refactor scidub CLI into different components
