+++ crates/scidub-cli/src/cli/init.rs
@@ -1,0 +1,103 @@
+use std::path::PathBuf;
+
+use clap::Parser;
+
+use crate::{
+    project::{Manifest, Project},
+    walkdir::RelWalkDir,
+};
+
+const INIT_GIT_IGNORE: &str = "\
+# Generated by `scidub init`
+
+# Ensure that the default `build/` directory is not tracked. This directory
+# is used to store files for building the project.
+/build/
+
+# The `game_files` directory is where the original game files are stored.
+# It is intentionally untracked to avoid committing potentially copyrighted
+# content into the project.
+/game_files/
+
+# Ensure the directory is kept using the .gitkeep file.
+!/game_files/.gitkeep
+";
+
+const INIT_SCIDUB_TOML: &str = "
+# Generated by `scidub init`
+game_data = \"game_data\"
+";
+
+/// Creates a new project.
+///
+/// This generates an initial project configuration, directory structure,
+/// and .gitignore file.
+///
+/// If a path to the game data is provided, it will also generate an initial
+/// manifest file for the game resources.
+#[derive(Debug, Parser)]
+pub(super) struct Init {
+    /// The root directory for the new project.
+    ///
+    /// If not specified, the current working directory is used.
+    #[clap(index = 1)]
+    root: Option<PathBuf>,
+
+    /// A path to the game data files to be copied into the project.
+    ///
+    /// If provided, the files will be copied to the `game_data/` directory
+    /// under the project root. By default, this directory will not be
+    /// tracked by Git (or other tools that observe .gitignore) to avoid
+    /// checking in potentially copyrighted files into a Git project.
+    #[clap(long)]
+    game_data: Option<PathBuf>,
+}
+
+impl Init {
+    pub(super) fn run(self) -> anyhow::Result<()> {
+        let root = self.root.map_or_else(std::env::current_dir, Ok)?;
+
+        anyhow::ensure!(root.exists(), "Project directory does not exist.");
+        anyhow::ensure!(
+            std::fs::read_dir(&root)?.next().is_none(),
+            "Project directory is not empty."
+        );
+
+        eprintln!("Creating project under {}", root.display());
+        let git_ignore_path = root.join(".gitignore");
+        let project_config_path = root.join("scidub.toml");
+        if !git_ignore_path.exists() {
+            std::fs::write(&git_ignore_path, INIT_GIT_IGNORE)?;
+        }
+
+        if !project_config_path.exists() {
+            std::fs::write(&git_ignore_path, INIT_SCIDUB_TOML)?;
+        }
+
+        let game_data_path = root.join("game_data");
+        std::fs::create_dir_all(&game_data_path)?;
+        std::fs::write(game_data_path.join(".gitkeep"), "")?;
+
+        if let Some(source_data) = &self.game_data {
+            for path in RelWalkDir::new(source_data) {
+                let path = path?;
+                let source_path = source_data.join(&path);
+                let dest_path = game_data_path.join(&path);
+                if let Some(dest_parent) = dest_path.parent() {
+                    std::fs::create_dir_all(dest_parent)?;
+                    std::fs::copy(&source_path, &dest_path)?;
+                }
+            }
+        }
+
+        let manifest = Manifest::generate_from_game_dir(&game_data_path)?;
+
+        // We are in a state where we can load the project directly.
+        let project = Project::new(root);
+
+        let mut output = std::fs::File::create(project.manifest_path()?)?;
+        serde_json::to_writer_pretty(&mut output, &manifest)?;
+
+        Ok(())
+    }
+}
