+++ crates/scidub-cli/src/main.rs
@@ -5,6 +5,7 @@ mod commands;
 mod data;
 mod dist_env;
 mod project;
+mod walkdir;
 
 fn main() {
     let cli = cli::Cli::parse();
